fix,feat: refatora listagem dos filmes e adiciona tooltip
