+++ frontend/package.json
@@ -19,6 +19,7 @@
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
+    "react-tooltip": "^5.18.0",
     "redux-thunk": "^2.4.2",
     "styled-components": "^6.0.1",
     "tailwindcss": "^3.3.2",
